refactor(app): componentize chat form
