+++ app/retrieval_agents/page.tsx
@@ -1,70 +1,30 @@
 import { ChatWindow } from "@/components/ChatWindow";
+import { UploadDocumentsForm } from "@/components/UploadDocumentsForm";
 
+/**
+  # 🦜 Retrieval Agent
+
+  - 🤝 Showcases a LangChain retrieval chain.
+  - 🛠️ The agent has access to a vector store retriever as a tool as well as a memory. It's particularly well suited to meta-questions about the current conversation.
+  - 🤖 By default, the agent is pretending to be a robot, but you can change the prompt to whatever you want!
+  - 🔱 Before running this example, you'll first need to set up a Supabase (or other) vector store. See the README for more details.
+  - 👇 Upload some text, then try asking e.g. `What are some ways of doing retrieval in LangChain?` below!
+ */
 export default function AgentsPage() {
-  const InfoCard = (
-    <div className="p-4 md:p-8 rounded w-full max-h-[85%] overflow-hidden">
-      <h1 className="text-3xl md:text-4xl mb-4">
-        LangChain.js Retrieval Agent 🦜🔗
-      </h1>
-      <ul>
-        <li className="hidden text-l md:block">
-          🤝
-          <span className="ml-2">
-            This template showcases a{" "}
-            <a href="https://js.langchain.com/" target="_blank">
-              LangChain.js
-            </a>{" "}
-            retrieval chain and the Vercel{" "}
-            <a href="https://sdk.vercel.ai/docs" target="_blank">
-              AI SDK
-            </a>{" "}
-            .
-          </span>
-        </li>
-        <li className="hidden text-l md:block">
-          🛠️
-          <span className="ml-2">
-            The agent has access to a vector store retriever as a tool as well
-            as a memory. It&apos;s particularly well suited to meta-questions
-            about the current conversation.
-          </span>
-        </li>
-        <li>
-          🤖
-          <span className="ml-2">
-            By default, the agent is pretending to be a robot, but you can
-            change the prompt to whatever you want!
-          </span>
-        </li>
-        <li className="hidden text-l md:block">
-          🔱
-          <span className="ml-2">
-            Before running this example, you&apos;ll first need to set up a
-            Supabase (or other) vector store. See the README for more details.
-          </span>
-        </li>
-        <li className="text-l">
-          👇
-          <span className="ml-2">
-            Upload some text, then try asking e.g.{" "}
-            <code>What are some ways of doing retrieval in LangChain?</code>{" "}
-            below!
-          </span>
-        </li>
-      </ul>
-    </div>
-  );
   return (
-    <ChatWindow
-      endpoint="api/chat/retrieval_agents"
-      emptyStateComponent={InfoCard}
-      showIngestForm={true}
-      showIntermediateStepsToggle={true}
-      placeholder={
-        'Beep boop! I\'m a robot retrieval-focused agent! Ask, "What are some ways of doing retrieval in LangChain.js?"'
-      }
-      emoji="🤖"
-      titleText="Robbie the Retrieval Robot"
-    ></ChatWindow>
+    <>
+      <UploadDocumentsForm />
+      <ChatWindow
+        endpoint="api/chat/retrieval_agents"
+        emptyStateComponent={<div />}
+        showIngestForm={true}
+        showIntermediateStepsToggle={true}
+        placeholder={
+          'Beep boop! I\'m a robot retrieval-focused agent! Ask, "What are some ways of doing retrieval in LangChain.js?"'
+        }
+        emoji="🤖"
+        titleText="Robbie the Retrieval Robot"
+      ></ChatWindow>
+    </>
   );
 }
